refine the four parts
